abstracted plants endpoints and model
