+++ auth/auth-router.js
@@ -57,103 +57,4 @@ router.post("/login", async (req, res, next) => {
   }
 });
 
-router.put("/users/:id", restrict, async (req, res, next) => {
-  // updates user info and returns newly updated user
-  try {
-    const payload = {
-      username: req.body.username,
-      password: await bcrypt.hash(req.body.password, 10),
-      phoneNumber: req.body.phoneNumber,
-    };
-
-    const updatedUser = await usersModel.updateUser(payload, req.params.id);
-
-    res.status(200).json(updatedUser);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.get("/users", async (req, res, next) => {
-  // returns list of users
-  try {
-    const users = await usersModel.getUsers();
-
-    res.status(200).json(users);
-  } catch (err) {
-    next(err);
-  }
-});
-
-// PLANT CRUD ENDPOINTS BELOW
-
-router.get("/plants/:id", restrict, async (req, res, next) => {
-  // returns list of plants for specified user
-  try {
-    const plants = await usersModel.getUserPlants(req.params.id);
-
-    res.status(200).json(plants);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.post("/plants", restrict, async (req, res, next) => {
-  // creates new plant and returns newly created plant
-  try {
-    const newplant = await usersModel.addPlant(req.body);
-
-    res.status(201).json(newplant);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.put("/plants/:id", restrict, async (req, res, next) => {
-  // updates plant and returns newly updated plant
-  try {
-    const updatedPlant = await usersModel.updatePlant(req.body, req.params.id);
-
-    res.status(200).json(updatedPlant);
-  } catch (err) {
-    next(err);
-  }
-});
-
-router.delete("/plants/:id", restrict, async (req, res, next) => {
-  // deletes plant returns deleted plant object
-  try {
-    const deletedPlant = await usersModel.removePlant(req.params.id);
-
-    res.status(200).json(deletedPlant);
-  } catch (err) {
-    next(err);
-  }
-});
-
-// RESTRICT MIDDELWARE THAT CHECKS IF THERE IS A JSON WEB TOKEN IN THE HEADERS OF THE REQUEST
-
-function restrict(req, res, next) {
-  try {
-    // verify that there is a request header named authorization
-    const token = req.headers.authorization;
-
-    // if token is not present return this
-    if (!token) {
-      return res.status(401).json({ message: "You are not authorized!" });
-    }
-
-    // verify the token to make sure it has not been tampered with
-    jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
-      if (err) {
-        return res.status(401).json({ message: "You are not authorized!" });
-      }
-
-      next();
-    });
-  } catch (err) {
-    next(err);
-  }
-}
-
 module.exports = router;
