+++ users/users-router.js
@@ -1,6 +1,9 @@
 const router = require("express").Router();
 const usersModel = require("./users-model");
+const plantsRouter = require("../plants/plants-router");
 const restrict = require("../middleware/restrict");
+
+router.use("/plants", plantsRouter);
 
 router.get("/", restrict, async (req, res) => {
   try {
